backend/API/requirements.txt
flask==3.0.0
flask-sqlalchemy==3.0.5
flask-cors==4.0.0
python-dotenv==1.0.0
openai==0.28.0
tensorflow==2.12.0
pillow==10.0.0